==> analysis/R_analysis/scr/00_libraries.R <==
renv::install("ica")
renv::install("mgcv")
renv::install("parallel")
renv::install("data.table")
renv::install("pryr")
renv::install("beeswarm")
renv::install("Rdpack")
renv::install("R.rsp")
renv::install("tidyverse")
renv::install("devtools")
renv::install("Seurat")
renv::install("ROCR")
renv::install("RColorBrewer")
renv::install("viridis")
renv::install("ggpubr")
renv::install("vioplot")
renv::install("devtools")

renv::install("bioc::limma")
renv::install("bioc::GEOquery")
renv::install("bioc::biomaRt")
renv::install("bioc::AnnotationHub")
renv::install("bioc::AnnotationDbi")
renv::install("bioc::ensembldb")

renv::install("bioc::affy")
renv::install("bioc::edgeR")

renv::install("LBMC/RAPToR")

# renv::install("mcp.project")
renv::install("LBMC/mouseRef")
renv::install("LBMC/zebraRef")
renv::install("LBMC/drosoRef")

# configure.args = c(build_vignettes = T)
# options(configure.args = configure.args)
renv::install("LBMC/wormRef")
renv::install("sf")
